+++ src/shared/components/Studio/ExpandableStudioList.less
@@ -20,10 +20,6 @@
     padding-top: 45px;
   }
 
-  .workspace-list {
-    margin-left: 24px;
-  }
-
   .studio-name,
   .workspace-title {
     color: @link-color;
